requirements.txt
# Python Dependencies for Plant Disease Detection ML API
# Using Keras with JAX backend for Hugging Face model
# Install with: pip install -r requirements.txt

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Image Processing
Pillow>=9.0.0

# Deep Learning - Keras (framework agnostic)
keras>=3.0.0

# JAX Backend (faster than TensorFlow)
jax>=0.4.0
jaxlib>=0.4.0

# ML Utilities
numpy>=1.20.0

# Utilities
python-dotenv>=1.0.0
huggingface-hub>=0.17.0
